requirements.txt
autopep8==1.4.4
boto3==1.10.42
botocore==1.13.42
Django==2.2.8
django-crispy-forms==1.8.1
docutils==0.15.2
jmespath==0.9.4
Pillow==6.2.1
pycodestyle==2.5.0
pynamodb==4.2.0
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
django-modeltranslation
psycopg2
gunicorn
django-heroku
django-better-admin-arrayfield
django-ckeditor
django-storages
django-imagekit
django-baton
